[SPARK-7462] By default retain group by columns in aggregate
